+++ src/main.ts
@@ -1,7 +1,7 @@
 export { createIndex, indexToObject, createIndexFromObject } from "@src/frontend/indexing";
 export { search } from "@src/frontend/search";
 export { LinearIndex } from "@src/method/linearindex";
-export { HyblidBigramInvertedIndex } from "@src/frontend/indextypes";
+export { HybridBigramInvertedIndex } from "@src/frontend/indextypes";
 export { UniSearchError } from "@src/frontend/base";
 export type {
     UniIndex,